Only Home Page Working
This is first version with minimum design, I just worked on learning how to create a responsive web page by just using HTML5 and CSS. This has minimum design, however i did try to use a color scheme of Crimson, Teal and Black. I still have a long way to go in learning, but this definitely is a start for me, and feels great.